requirements.txt
opencv-python
dlib
numpy
pyautogui
Pillow
